Move previewEntries to internal_binding.
