+++ ext/node/polyfills/internal/console/constructor.mjs
@@ -17,17 +17,7 @@ import {
   validateInteger,
   validateObject,
 } from "ext:deno_node/internal/validators.mjs";
-const previewEntries = (iter, isKeyValue) => {
-  if (isKeyValue) {
-    const arr = [...iter];
-    if (Array.isArray(arr[0]) && arr[0].length === 2) {
-      return [[].concat(...arr), true];
-    }
-    return [arr, false];
-  } else {
-    return [...iter];
-  }
-};
+import { previewEntries } from "ext:deno_node/internal_binding/util.ts";
 import { Buffer } from "node:buffer";
 const { isBuffer } = Buffer;
 import {
